Change how databases are handled
This is how it works now:
- Load in a clean database dump at the beginning. Which file to load
  can be set in the config file
- Let the user choose to dump the database into a file with a name that
  contains the bugnumber and date/time (so they can go back to it if
  they want to)

This means you have to make sure you have a clean database to test on
in the first place.

Future enhancement: add an option to the script to load update and
dump the database with the standard filename.

Huge thanks to Liz Rea for suggesting these changes!